+++ lib/VovkPLCRuntime/src/tools/arithmetics/methods-flow.h
@@ -24,7 +24,7 @@
 
 
 
-RuntimeError printOpcodeAt(const uint8_t* program, uint16_t size, uint16_t index) {
+RuntimeError printOpcodeAt(const uint8_t* program, uint32_t size, uint32_t index) {
     if (index >= size) return INVALID_PROGRAM_INDEX;
     PLCRuntimeInstructionSet opcode = (PLCRuntimeInstructionSet) program[index];
     bool valid_opcode = OPCODE_EXISTS(opcode);
@@ -48,7 +48,7 @@ RuntimeError printOpcodeAt(const uint8_t* program, uint16_t size, uint16_t index
     Serial.print(']');
     return STATUS_SUCCESS;
 }
-RuntimeError printlnOpcodeAt(const uint8_t* program, uint16_t size, uint16_t index) {
+RuntimeError printlnOpcodeAt(const uint8_t* program, uint32_t size, uint32_t index) {
     RuntimeError error = printOpcodeAt(program, size, index);
     Serial.println();
     return error;
@@ -71,9 +71,9 @@ RuntimeError printlnOpcodeAt(const uint8_t* program, uint16_t size, uint16_t ind
 
 namespace PLCMethods {
 
-    RuntimeError handle_JMP(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
-        uint16_t size = 2;
+    RuntimeError handle_JMP(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
+        uint32_t size = 2;
         if (index + size > program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         u8A_to_u16 cvt;
         cvt.u8A[1] = program[index];
@@ -82,9 +82,9 @@ namespace PLCMethods {
         if (index >= program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         return STATUS_SUCCESS;
     }
-    RuntimeError handle_JMP_IF(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
-        uint16_t size = 2;
+    RuntimeError handle_JMP_IF(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
+        uint32_t size = 2;
         if (index + size > program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         if (stack->pop_bool()) {
             u8A_to_u16 cvt;
@@ -95,9 +95,9 @@ namespace PLCMethods {
         } else index += size;
         return STATUS_SUCCESS;
     }
-    RuntimeError handle_JMP_IF_NOT(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
-        uint16_t size = 2;
+    RuntimeError handle_JMP_IF_NOT(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
+        uint32_t size = 2;
         if (index + size > program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         if (!stack->pop_bool()) {
             u8A_to_u16 cvt;
@@ -109,9 +109,9 @@ namespace PLCMethods {
         return STATUS_SUCCESS;
     }
 
-    RuntimeError handle_CALL(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
-        uint16_t size = 2;
+    RuntimeError handle_CALL(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
+        uint32_t size = 2;
         if (index + size > program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         u8A_to_u16 cvt;
         cvt.u8A[1] = program[index];
@@ -123,9 +123,9 @@ namespace PLCMethods {
         if (index >= program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         return STATUS_SUCCESS;
     }
-    RuntimeError handle_CALL_IF(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
-        uint16_t size = 2;
+    RuntimeError handle_CALL_IF(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
+        uint32_t size = 2;
         if (index + size > program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         if (stack->pop_bool()) {
             u8A_to_u16 cvt;
@@ -139,9 +139,9 @@ namespace PLCMethods {
         } else index += size;
         return STATUS_SUCCESS;
     }
-    RuntimeError handle_CALL_IF_NOT(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
-        uint16_t size = 2;
+    RuntimeError handle_CALL_IF_NOT(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
+        uint32_t size = 2;
         if (index + size > program_size) CHECK_PROGRAM_POINTER_BOUNDS_HEAD;
         if (!stack->pop_bool()) {
             u8A_to_u16 cvt;
@@ -155,8 +155,8 @@ namespace PLCMethods {
         } else index += size;
         return STATUS_SUCCESS;
     }
-    RuntimeError handle_RET(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
-        IGNORE_UNUSED uint16_t index_start = index;
+    RuntimeError handle_RET(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
+        IGNORE_UNUSED uint32_t index_start = index;
         if (stack->call_stack->size() == 0) return CALL_STACK_UNDERFLOW;
         uint16_t ret_index = stack->popCall();
         index = ret_index;
@@ -166,7 +166,7 @@ namespace PLCMethods {
     }
 
 
-    RuntimeError handle_EXIT(RuntimeStack* stack, uint8_t* program, uint16_t program_size, uint16_t& index) {
+    RuntimeError handle_EXIT(RuntimeStack* stack, uint8_t* program, uint32_t program_size, uint32_t& index) {
         if (index >= program_size) return STATUS_SUCCESS;
         uint8_t exit_code = program[index++];
         return (RuntimeError) exit_code;
